Fix bug where xform fields that were not specified for motion interpolation were still getting interpolated.
Carry logic from this fix to writing Xform Xmls as well.

Motion was not supported on some Xform fields as it was originally copied from flam3. Allow it on all but xaos now.

Fix clamping range for m_ColorSpeed at the end of ApplyMotion() to match the values we accept elsewhere.